+++ packages/cdn/CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Added
+ - `cdn_organization.OKPOCode` will be trimmed during insert/update operation (prevents unwonted whitespaces remove whitespaces at start/end)
 
 ### Changed